- Simplified RootLayoutContent - Removed unused code in RouteHandler - Added RouteHandler for routing and tab management
